+++ entrypoint.sh
@@ -3,4 +3,4 @@
 # This is the entrypoint for the Switchboard Server docker image.
 # The purpose of this script is to aid in passing command line arguments, based on the environment variables.
 
-/app/bin/switchboard --botpsk "$SB_BOTPSK" --token "$BOT_TOKEN" --docker
+/app/bin/switchboard --token "$BOT_TOKEN" --docker